Move bits to the right places.
